available {rversions}	R Documentation

Available R Builds

Description

Available R Builds

Usage

available(platform = current_r_platform())

Arguments

platform
Operating system version. Defaults to current_r_platform(). Possible values are:

windows: x86_64 Windows.

windows-x86_64: the same.

windows-aarch64: aarch64 Windows.

macos: arm64 macOS.

macos-arm64: the same.

macos-x86_64: x86_64 macOS.

linux-<distro>-<release>: x86_64 Linux on the given distribution and release, e.g. linux-ubuntu-22.04.

A target triplet of the form cpu-vendor-os-abi. On Linux it should also include the distribution name and release. See current_r_platform() for examples.

Details

available() caches its results in the current R session. To clear the cache restart R or set the R_VERSION_CACHE_TIMEOUT environment variable to 0.

Value

A data frame with columns version, date, semver, type, and url.

Examples

available()
available("windows")
available("macos-arm64")
available("linux-gnu-ubuntu-24.04")
available("aarch64-unknown-linux-gnu-ubuntu-22.04")

[Package rversions version 3.0.0 Index]