current_r_platform {rversions}	R Documentation

Current R platform

Description

current_r_platform() detects the platform of the current R version.

Usage

current_r_platform()

Details

It returns a string of the form cpu-vendor-os-abi, a target triplet, see some examples below. On Linux it also includes the distribution name and version and part of the ABI.

Examples:

x86_64-apple-darwin17.0: macOS High Sierra.

aarch64-apple-darwin20: macOS Big Sur on arm64.

x86_64-w64-mingw32: 64 bit Windows.

i386-w64-mingw32: 32 bit Windows.

i386+x86_64-w64-mingw32: 64 bit + 32 bit Windows.

i386-pc-solaris2.10: 32 bit Solaris. (Some broken 64 Solaris builds might have the same platform string, unfortunately.)

x86_64-pc-linux-gnu-debian-10: Debian Linux 10 on x86_64.

x86_64-pc-linux-musl-alpine-3.14.1: Alpine Linux.

x86_64-pc-linux-gnu-unknown: Unknown Linux Distribution on x86_64.

s390x-ibm-linux-gnu-ubuntu-20.04: Ubuntu Linux 20.04 on S390x.

amd64-portbld-freebsd12.1: FreeBSD 12.1 on x86_64.

Examples

current_r_platform()

[Package rversions version 3.0.0 Index]