linux_distros {rversions}	R Documentation

Return information about Posit's (and some other) R builds for Linux

Description

Return information about Posit's (and some other) R builds for Linux

Usage

linux_distros()

Details

linux_distros() caches its results in the current R session. To clear the cache restart R or set the R_VERSION_CACHE_TIMEOUT environment variable to 0.

Value

A data frame with the following columns:

id: The distro ID, e.g. ubuntu-2404.

aliases: A list column of character vectors with alternative names for the distro. Alternative names starting and ending with a forward slash are regular expressions.

name: The full name of the distro, e.g. Red Hat Enterprise Linux.

version: The version of the distro, e.g. 24.04 for Ubuntu, or 9 for RHEL.

codename: The codename of the distro, e.g. focal for Ubuntu 20.04.

url: The URL of the distro's homepage.

docker: The name of the official Docker image for the distro, e.g. ubuntu:20.04.

eol: The end-of-life date of the distro, or NA if unknown.

retired: whether the distro version is retired for Posit and R-hub.

aarch64: whether aarch64 builds are available for this distro. Possible values are posit (builds are available from Posit), r-hub (builds are available from R-hub), none (no builds are available).

ppm_binaries: whether PPM binaries are available for this distro.

ppm_aarch64: whether aarch64 PPM binaries are available for this distro.

ppm_url: The URL of the PPM binaries for this distro.

Examples

linux_distros()

[Package rversions version 3.0.0 Index]