r_oldrel {rversions}	R Documentation

Version number of R-oldrel

Description

R-oldrel is the latest version of the previous minor version. We extract version numbers from the R SVN repository tags.

Usage

r_oldrel(dots = TRUE)

Arguments

dots
Whether to use dots instead of dashes in the version number.

Value

A one row data frame, with columns ‘version’, ‘date’ and ‘nickname’.

Note

This function is superseded by resolve():

r_oldrel()
#> # A data frame: 1 x 5
#>   version date                nickname    semver     URL                        
#>   <chr>   <dttm>              <chr>       <pckg_vrs> <chr>                      
#> 1 4.4.3   2025-02-28 08:08:59 Trophy Case 4.4.3      https://cran.rstudio.com/s~

resolve("oldrel", platform = NA)
#> # A data frame: 1 x 6
#>   version date                nickname    semver     url                   type 
#>   <chr>   <dttm>              <chr>       <pckg_vrs> <chr>                 <chr>
#> 1 4.4.3   2025-02-28 08:08:59 Trophy Case 4.4.3      https://cran.rstudio~ oldr~

Examples

r_oldrel()

[Package rversions version 3.0.0 Index]