r_release_macos {rversions}	R Documentation

Latest R release with a macOS installer available for download

Description

r_release() works based on the SVN tags in the main R source code repository. However, an SVN tag does not mean that the same version is available for download. r_release_macos returns the latest version for which a macOS installer is available.

Usage

r_release_macos(arch = c("x86_64", "arm64"))

Arguments

arch
The architecture, either ‘x86_64’ or ‘arm64’.

Value

A one row data frame, with columns ‘version’, ‘date’, ‘URL’ and ‘nickname’.

Note

This function is superseded by resolve():

r_release_macos()
#> # A data frame: 1 x 5
#>   version date                nickname          semver     URL                  
#>   <chr>   <dttm>              <chr>             <pckg_vrs> <chr>                
#> 1 4.5.1   2025-06-13 07:12:01 Great Square Root 4.5.1      https://cran.rstudio~

resolve("release", platform = "macos")
#> # A data frame: 1 x 6
#>   version date                nickname          semver     url             type 
#>   <chr>   <dttm>              <chr>             <pckg_vrs> <chr>           <chr>
#> 1 4.5.1   2025-06-13 07:12:01 Great Square Root 4.5.1      https://cran.r~ rele~

See Also

Other R downloads: r_release_tarball(), r_release_win()

Examples

r_release()
r_release_macos()
r_release_macos(arch = "arm64")

[Package rversions version 3.0.0 Index]