r_release_win {rversions}	R Documentation

Latest R release with a Windows installer available for download

Description

r_release() works based on the SVN tags in the main R source code repository. However, an SVN tag does not mean that the same version is available for download. r_release_win returns the latest version for which a Windows installer is available.

Usage

r_release_win()

Value

A one row data frame, with columns ‘version’, ‘date’, ‘URL’ and ‘nickname’.

Note

This function is superseded by resolve():

r_release_win()
#> # A data frame: 1 x 5
#>   version date                nickname          semver     URL                  
#>   <chr>   <dttm>              <chr>             <pckg_vrs> <chr>                
#> 1 4.5.1   2025-06-13 07:12:01 Great Square Root 4.5.1      https://cran.rstudio~

resolve("release", platform = "windows")
#> # A data frame: 1 x 7
#>   version date                nickname          semver    url   type  rtools_url
#>   <chr>   <dttm>              <chr>             <pckg_vr> <chr> <chr> <chr>     
#> 1 4.5.1   2025-06-13 07:12:01 Great Square Root 4.5.1     http~ rele~ https://g~

See Also

Other R downloads: r_release_macos(), r_release_tarball()

Examples

r_release()
r_release_win()

[Package rversions version 3.0.0 Index]