Available R Builds

Available R Builds

available(platform = current_r_platform())

Arguments

platform

Operating system version. Defaults to current_r_platform(). Possible values are:

windows: x86_64 Windows.

windows-x86_64: the same.

windows-aarch64: aarch64 Windows.

macos: arm64 macOS.

macos-arm64: the same.

macos-x86_64: x86_64 macOS.

linux-<distro>-<release>: x86_64 Linux on the given distribution and release, e.g. linux-ubuntu-22.04.

A target triplet of the form cpu-vendor-os-abi. On Linux it should also include the distribution name and release. See current_r_platform() for examples.

Value

A data frame with columns version, date, semver, type, and url.

Details

available() caches its results in the current R session. To clear the cache restart R or set the R_VERSION_CACHE_TIMEOUT environment variable to 0.

Examples

available()
#> # A data frame: 57 × 5
#>    version date                semver     type    url                           
#>    <chr>   <dttm>              <pckg_vrs> <chr>   <chr>                         
#>  1 3.0.0   2013-04-03 07:12:36 3.0.0      release https://cdn.posit.co/r/ubuntu…
#>  2 3.0.1   2013-05-16 07:11:33 3.0.1      release https://cdn.posit.co/r/ubuntu…
#>  3 3.0.2   2013-09-25 07:11:09 3.0.2      release https://cdn.posit.co/r/ubuntu…
#>  4 3.0.3   2014-03-06 08:12:33 3.0.3      release https://cdn.posit.co/r/ubuntu…
#>  5 3.1.0   2014-04-10 07:11:10 3.1.0      release https://cdn.posit.co/r/ubuntu…
#>  6 3.1.1   2014-07-10 07:11:09 3.1.1      release https://cdn.posit.co/r/ubuntu…
#>  7 3.1.2   2014-10-31 08:11:32 3.1.2      release https://cdn.posit.co/r/ubuntu…
#>  8 3.1.3   2015-03-09 08:12:20 3.1.3      release https://cdn.posit.co/r/ubuntu…
#>  9 3.2.0   2015-04-16 07:13:33 3.2.0      release https://cdn.posit.co/r/ubuntu…
#> 10 3.2.1   2015-06-18 07:15:04 3.2.1      release https://cdn.posit.co/r/ubuntu…
#> # ℹ 47 more rows
available("windows")
#> # A data frame: 97 × 5
#>    version date                semver     type    url                           
#>    <chr>   <dttm>              <pckg_vrs> <chr>   <chr>                         
#>  1 2.0.0   2004-10-04 14:24:38 2.0.0      release https://cran-archive.r-projec…
#>  2 2.0.1   2004-11-15 14:16:30 2.0.1      release https://cran-archive.r-projec…
#>  3 2.1.0   2005-04-18 22:26:33 2.1.0      release https://cran-archive.r-projec…
#>  4 2.1.1   2005-06-20 09:27:13 2.1.1      release https://cran-archive.r-projec…
#>  5 2.2.0   2005-10-06 10:22:14 2.2.0      release https://cran-archive.r-projec…
#>  6 2.2.1   2005-12-20 10:35:21 2.2.1      release https://cran-archive.r-projec…
#>  7 2.3.0   2006-04-24 10:37:20 2.3.0      release https://cran-archive.r-projec…
#>  8 2.3.1   2006-06-01 08:25:33 2.3.1      release https://cran-archive.r-projec…
#>  9 2.4.0   2006-10-03 10:15:04 2.4.0      release https://cran-archive.r-projec…
#> 10 2.4.1   2006-12-18 09:49:23 2.4.1      release https://cran-archive.r-projec…
#> # ℹ 87 more rows
available("macos-arm64")
#> # A data frame: 20 × 5
#>    version date                semver     type    url                           
#>    <chr>   <dttm>              <pckg_vrs> <chr>   <chr>                         
#>  1 4.1.0   2021-05-18 07:05:22 4.1.0      release https://cran.rstudio.com/bin/…
#>  2 4.1.1   2021-08-10 07:05:06 4.1.1      release https://cran.rstudio.com/bin/…
#>  3 4.1.2   2021-11-01 08:05:12 4.1.2      release https://cran.rstudio.com/bin/…
#>  4 4.1.3   2022-03-10 08:05:38 4.1.3      release https://cran.rstudio.com/bin/…
#>  5 4.2.0   2022-04-22 07:05:41 4.2.0      release https://cran.rstudio.com/bin/…
#>  6 4.2.1   2022-06-23 07:05:33 4.2.1      release https://cran.rstudio.com/bin/…
#>  7 4.2.2   2022-10-31 08:05:54 4.2.2      release https://cran.rstudio.com/bin/…
#>  8 4.2.3   2023-03-15 08:06:01 4.2.3      release https://cran.rstudio.com/bin/…
#>  9 4.3.0   2023-04-21 07:06:14 4.3.0      release https://cran.rstudio.com/bin/…
#> 10 4.3.1   2023-06-16 07:06:07 4.3.1      release https://cran.rstudio.com/bin/…
#> 11 4.3.2   2023-10-31 08:07:42 4.3.2      release https://cran.rstudio.com/bin/…
#> 12 4.3.3   2024-02-29 08:07:53 4.3.3      release https://cran.rstudio.com/bin/…
#> 13 4.4.0   2024-04-24 04:07:56 4.4.0      release https://cran.rstudio.com/bin/…
#> 14 4.4.1   2024-06-14 07:08:17 4.4.1      release https://cran.rstudio.com/bin/…
#> 15 4.4.2   2024-10-31 08:09:02 4.4.2      release https://cran.rstudio.com/bin/…
#> 16 4.4.3   2025-02-28 08:08:59 4.4.3      release https://cran.rstudio.com/bin/…
#> 17 4.5.0   2025-04-11 08:51:39 4.5.0      release https://cran.rstudio.com/bin/…
#> 18 4.5.1   2025-06-13 07:12:01 4.5.1      release https://cran.rstudio.com/bin/…
#> 19 4.5.1   NA                  4.5.1      patched https://mac.r-project.org/big…
#> 20 4.6.0   NA                  4.6.0      devel   https://mac.r-project.org/big…
available("linux-gnu-ubuntu-24.04")
#> # A data frame: 57 × 5
#>    version date                semver     type    url                           
#>    <chr>   <dttm>              <pckg_vrs> <chr>   <chr>                         
#>  1 3.0.0   2013-04-03 07:12:36 3.0.0      release https://cdn.posit.co/r/ubuntu…
#>  2 3.0.1   2013-05-16 07:11:33 3.0.1      release https://cdn.posit.co/r/ubuntu…
#>  3 3.0.2   2013-09-25 07:11:09 3.0.2      release https://cdn.posit.co/r/ubuntu…
#>  4 3.0.3   2014-03-06 08:12:33 3.0.3      release https://cdn.posit.co/r/ubuntu…
#>  5 3.1.0   2014-04-10 07:11:10 3.1.0      release https://cdn.posit.co/r/ubuntu…
#>  6 3.1.1   2014-07-10 07:11:09 3.1.1      release https://cdn.posit.co/r/ubuntu…
#>  7 3.1.2   2014-10-31 08:11:32 3.1.2      release https://cdn.posit.co/r/ubuntu…
#>  8 3.1.3   2015-03-09 08:12:20 3.1.3      release https://cdn.posit.co/r/ubuntu…
#>  9 3.2.0   2015-04-16 07:13:33 3.2.0      release https://cdn.posit.co/r/ubuntu…
#> 10 3.2.1   2015-06-18 07:15:04 3.2.1      release https://cdn.posit.co/r/ubuntu…
#> # ℹ 47 more rows
available("aarch64-unknown-linux-gnu-ubuntu-22.04")
#> # A data frame: 57 × 5
#>    version date                semver     type    url                           
#>    <chr>   <dttm>              <pckg_vrs> <chr>   <chr>                         
#>  1 3.0.0   2013-04-03 07:12:36 3.0.0      release https://cdn.posit.co/r/ubuntu…
#>  2 3.0.1   2013-05-16 07:11:33 3.0.1      release https://cdn.posit.co/r/ubuntu…
#>  3 3.0.2   2013-09-25 07:11:09 3.0.2      release https://cdn.posit.co/r/ubuntu…
#>  4 3.0.3   2014-03-06 08:12:33 3.0.3      release https://cdn.posit.co/r/ubuntu…
#>  5 3.1.0   2014-04-10 07:11:10 3.1.0      release https://cdn.posit.co/r/ubuntu…
#>  6 3.1.1   2014-07-10 07:11:09 3.1.1      release https://cdn.posit.co/r/ubuntu…
#>  7 3.1.2   2014-10-31 08:11:32 3.1.2      release https://cdn.posit.co/r/ubuntu…
#>  8 3.1.3   2015-03-09 08:12:20 3.1.3      release https://cdn.posit.co/r/ubuntu…
#>  9 3.2.0   2015-04-16 07:13:33 3.2.0      release https://cdn.posit.co/r/ubuntu…
#> 10 3.2.1   2015-06-18 07:15:04 3.2.1      release https://cdn.posit.co/r/ubuntu…
#> # ℹ 47 more rows