Return information about Posit's (and some other) R builds for Linux

Return information about Posit's (and some other) R builds for Linux

linux_distros()

Value

A data frame with the following columns:

id: The distro ID, e.g. ubuntu-2404.

aliases: A list column of character vectors with alternative names for the distro. Alternative names starting and ending with a forward slash are regular expressions.

name: The full name of the distro, e.g. Red Hat Enterprise Linux.

version: The version of the distro, e.g. 24.04 for Ubuntu, or 9 for RHEL.

codename: The codename of the distro, e.g. focal for Ubuntu 20.04.

url: The URL of the distro's homepage.

docker: The name of the official Docker image for the distro, e.g. ubuntu:20.04.

eol: The end-of-life date of the distro, or NA if unknown.

retired: whether the distro version is retired for Posit and R-hub.

aarch64: whether aarch64 builds are available for this distro. Possible values are posit (builds are available from Posit), r-hub (builds are available from R-hub), none (no builds are available).

ppm_binaries: whether PPM binaries are available for this distro.

ppm_aarch64: whether aarch64 PPM binaries are available for this distro.

ppm_url: The URL of the PPM binaries for this distro.

Details

linux_distros() caches its results in the current R session. To clear the cache restart R or set the R_VERSION_CACHE_TIMEOUT environment variable to 0.

Examples

linux_distros()
#> # A data frame: 42 × 13
#>    id    aliases name  version codename url   docker eol                 retired
#>  * <chr> <I<nam> <chr> <chr>   <chr>    <chr> <chr>  <dttm>              <lgl>  
#>  1 alma… <chr>   Alma… 10      NA       http… almal… 2035-05-31 00:00:00 FALSE  
#>  2 alma… <NULL>  Alma… 8       NA       http… almal… 2029-03-01 00:00:00 FALSE  
#>  3 alma… <chr>   Alma… 9       NA       http… almal… 2032-05-31 00:00:00 FALSE  
#>  4 cent… <chr>   Cent… 6       NA       http… cento… 2020-11-30 00:00:00 TRUE   
#>  5 cent… <chr>   Cent… 7       NA       http… cento… 2024-06-30 00:00:00 FALSE  
#>  6 cent… <NULL>  Cent… 8       NA       http… cento… 2024-05-31 00:00:00 FALSE  
#>  7 debi… <chr>   Debi… 10      buster   http… debia… 2024-06-30 00:00:00 TRUE   
#>  8 debi… <chr>   Debi… 11      bullseye http… debia… 2024-08-14 00:00:00 TRUE   
#>  9 debi… <chr>   Debi… 12      bookworm http… debia… NA                  FALSE  
#> 10 debi… <chr>   Debi… 9       stretch  http… debia… 2022-06-30 00:00:00 TRUE   
#> # ℹ 32 more rows
#> # ℹ 4 more variables: aarch64 <chr>, ppm_binaries <lgl>, ppm_aarch64 <lgl>,
#> #   ppm_url <chr>