r_oldrel {rversions}	R Documentation

Version number of R-oldrel

Description

R-oldrel is the latest version of the previous minor version. We extract version numbers from the R SVN repository tags.

Usage

r_oldrel(dots = TRUE)

Arguments

dots
Whether to use dots instead of dashes in the version number.

Value

A one row data frame, with columns ‘version’, ‘date’ and ‘nickname’.

Examples

r_oldrel()

[Package rversions version 2.1.1 Index]