Latest release for which a source tarball is available for download

r_release() works based on the SVN tags in the main R source code repository. However, an SVN tag does not mean that the same version is available for download. r_release_tarball returns the latest version for which a source tarball is available.

r_release_tarball()

Value

A one row data frame, with columns ‘version’, ‘date’, ‘URL’ and ‘nickname’.

Note

This function is superseded by resolve():

r_release_tarball()
#> # A data frame: 1 x 5
#>   version date                nickname          semver     URL
#>   <chr>   <dttm>              <chr>             <pckg_vrs> <chr>
#> 1 4.5.1   2025-06-13 07:12:01 Great Square Root 4.5.1      https://cran.rstudio~

resolve("release", platform = NA)
#> # A data frame: 1 x 6
#>   version date                nickname          semver     url             type
#>   <chr>   <dttm>              <chr>             <pckg_vrs> <chr>           <chr>
#> 1 4.5.1   2025-06-13 07:12:01 Great Square Root 4.5.1      https://cran.r~ rele~

See also

Examples

r_release()
#> # A data frame: 1 × 6
#>   version date                nickname          semver     URL             type 
#>   <chr>   <dttm>              <chr>             <pckg_vrs> <chr>           <chr>
#> 1 4.5.1   2025-06-13 07:12:01 Great Square Root 4.5.1      https://cran.r… rele…
r_release_tarball()
#> # A data frame: 1 × 5
#>   version date                nickname          semver     URL                  
#>   <chr>   <dttm>              <chr>             <pckg_vrs> <chr>                
#> 1 4.5.1   2025-06-13 07:12:01 Great Square Root 4.5.1      https://cran.rstudio…