r_release_win {rversions}	R Documentation

Latest R release with a Windows installer available for download

Description

r_release() works based on the SVN tags in the main R source code repository. However, an SVN tag does not mean that the same version is available for download. r_release_win returns the latest version for which a Windows installer is available.

Usage

r_release_win()

Value

A one row data frame, with columns ‘version’, ‘date’, ‘URL’ and ‘nickname’.

See Also

Other R downloads: r_release_macos(), r_release_tarball()

Examples

r_release()
r_release_win()

[Package rversions version 2.1.1 Index]