r_versions {rversions}	R Documentation

Query R's past and present versions

Description

R version numbers consist of three numbers (since version 1.4.1): major, minor and patch.

Usage

r_versions(dots = TRUE)

Arguments

dots
Whether to use dots instead of dashes in the version number.

Details

We extract the version numbers from the tags in the SVN repository.

Value

A data frame with three columns: ‘version’, ‘date’ and ‘nickname’.

Examples

r_versions()

[Package rversions version 2.1.1 Index]