Resolve an R version specification

Resolve an R version specification

resolve(version = "release", platform = current_r_platform())

Arguments

version

Symbolic version name to resolve. Possible values are

release: the latest release version,

devel: Development version of R.

next: The next version of R. This is R-patched if there is no ongoing release process currently. Otherwise it may be R-alpha, R-beta, or R-rc or R-prerelease.

oldrel: The previous version of R, not counting patch versions.

oldrel/<n>: The n-th previous version of R, not counting patch versions.

x.y.z: A specific R version, e.g. 4.5.1.

x.y: The last patch release from a minor R branch, e.g. 4.4.

platform

Operating system version. Defaults to current_r_platform(). Possible values are:

NA: to get information about platform-independent release archives.

windows: x86_64 Windows.

windows-x86_64: the same.

windows-aarch64: aarch64 Windows.

macos: arm64 macOS.

macos-arm64: the same.

macos-x86_64: x86_64 macOS.

linux-<distro>-<release>: x86_64 Linux on the given distribution and release, e.g. linux-ubuntu-22.04.

A target triplet of the form cpu-vendor-os-abi. On Linux it should also include the distribution name and release. See current_r_platform() for examples.

Value

A one row data frame, with columns version, date, nickname, semver, url, type, rtools_url (only for Windows).

Details

resolve() caches its results in the current R session. To clear the cache restart R or set the R_VERSION_CACHE_TIMEOUT environment variable to 0.

Examples

resolve("release", "windows")
#> # A data frame: 1 × 7
#>   version date                nickname          semver    url   type  rtools_url
#>   <chr>   <dttm>              <chr>             <pckg_vr> <chr> <chr> <chr>     
#> 1 4.5.1   2025-06-13 07:12:01 Great Square Root 4.5.1     http… rele… https://g…
resolve("4.4", "windows-aarch64")
#> # A data frame: 1 × 7
#>   version date                nickname    semver     url        type  rtools_url
#>   <chr>   <dttm>              <chr>       <pckg_vrs> <chr>      <chr> <chr>     
#> 1 4.4.3   2025-02-28 08:08:59 Trophy Case 4.4.3      https://g… 4.4   https://g…
resolve("oldrel", "macos")
#> # A data frame: 1 × 6
#>   version date                nickname    semver     url                   type 
#>   <chr>   <dttm>              <chr>       <pckg_vrs> <chr>                 <chr>
#> 1 4.4.3   2025-02-28 08:08:59 Trophy Case 4.4.3      https://cran.rstudio… oldr…
resolve("devel", "linux-gnu-ubuntu-24.04")
#> # A data frame: 1 × 6
#>   version date   nickname                semver     url                    type 
#>   <chr>   <dttm> <chr>                   <pckg_vrs> <chr>                  <chr>
#> 1 4.6.0   NA     Unsuffered Consequences 4.6.0      https://cdn.posit.co/… devel
resolve("devel", "aarch64-unknown-linux-gnu-ubuntu-22.04")
#> # A data frame: 1 × 6
#>   version date   nickname                semver     url                    type 
#>   <chr>   <dttm> <chr>                   <pckg_vrs> <chr>                  <chr>
#> 1 4.6.0   NA     Unsuffered Consequences 4.6.0      https://cdn.posit.co/… devel